requirements.txt
streamlit==1.33.0
PyMuPDF==1.24.9
google-genai==0.3.0
protobuf==4.25.3